Hrse.jl Documentation

Hrse.jl Documentation

Getting Started

API

Getting Started

Hrse.jl is a Julia package for reading and writing files written in HRSE, or Human Readable S-Expressions, which are a human-readable format for representing data and configuration files that is equivalent and and interchangeable with s-expressions.

Hrse.jl provides two main functions, readhrse and writehrse, which read and write HRSE files, respectively.

API

Hrse.readhrse — Function

readhrse(hrse::IO; options=ReadOptions())
readhrse(hrse::String; options=ReadOptions())

Reads an HRSE file from the given IO object or string and returns the corresponding Julia object. The options argument can be used to configure the parser. Lists will be read as vectors, pairs as a Pair, symbols and strings as a String, and numeric types as the corresponding Julia type defined in the parser options.

Examples

julia> import Hrse

julia> hrse = """
       alpha:
           1 2 3 4
           5 6
           7 8 9
       beta: (0 . 3)
       gamma:
           a: 1
           b: 2
           c: "c"
       """;

julia> Hrse.readhrse(hrse)
3-element Vector{Pair{String}}:
 "alpha" => [[1, 2, 3, 4], [5, 6], [7, 8, 9]]
  "beta" => (0 => 3)
 "gamma" => Pair{String}["a" => 1, "b" => 2, "c" => "c"]

See also ReadOptions.

Hrse.ReadOptions — Type

ReadOptions(kwargs...)

Stores options for parsing HRSE files.

Arguments

integertypes = [Int64, BigInt]: A list of signed integer types to try parsing integers as. The first type that can represent the integer will be used.

floattype = Float64: The floating point type to parse floating point numbers as.

readcomments = false: Whether to read comments and store them in CommentedElement objects; if false, comments are ignored.

extensions: A collection of Extensions to HRSE.

See also readhrse.

Hrse.writehrse — Function

writehrse(io::IO, obj, options::PrinterOptions)
writehrse(obj, options::PrinterOptions)

Writes the given Julia object to the given IO object as a HRSE file. The options argument can be used to configure the behavior of the printer. If no IO object is given, the output is written to stdout.

Examples

julia> import Hrse

julia> hrse = [
           :alpha => [
               [1, 2, 3, 4],
               [5, 6],
               [7, 8, 9]
           ],
           :beta => (0 => 3),
           :gamma => [
               :a => 1
               :b => 2
               :c => :c
           ]
       ];

julia> Hrse.writehrse(hrse, Hrse.PrinterOptions())
alpha: 
  1 2 3 4
  5 6
  7 8 9
beta: 0: 3
gamma: 
  a: 1
  b: 2
  c: c

See also PrinterOptions.

Hrse.ashrse — Function

ashrse(obj, options::PrinterOptions)

Returns the given Julia object as a string containing a HRSE file. The options argument can be used to configure the behavior of the printer.

See also writehrse, PrinterOptions.

Hrse.PrinterOptions — Type

PrinterOptions(kwargs...)

Stores options for printing HRSE structures to text.

Arguments

indent = " ": The string to use for indentation.

comments = true: Whether to print comments from CommentedElement objects; if false, comments are ignored.

extensions: A collection of Extensions to HRSE.

pairmode = COLON_MODE: The PairMode to use when printing pairs.

inlineprimitives = 20: The maximum string length of a list of primitives to print on a single line instead of adding a new indentation level.

See also writehrse, ashrse.

Hrse.CommentedElement — Type

CommentedElement(element, comments)

Wraps an element with a list of comments which directly precede it attached to it.

Hrse.Extension — Type

An extension to HRSE's syntax that affects parsing and/or printing.

See also DENSE.

Hrse.DENSE — Constant

An extension to HRSE; axpects a "dense" HRSE file with a single root-level element instead of a list of elements.

See also Extension.

Hrse.PairMode — Type

A flag that decides how pairs are displayed while printing a HRSE structure.

See also CONDENSED_MODE, DOT_MODE, EQUALS_MODE, COLON_MODE.

Hrse.CONDENSED_MODE — Constant

A pair printing mode; pairs should all be condensed to a single line using classic s-expressions, eleminating as many spaces as possible.

See also PairMode.

Hrse.DOT_MODE — Constant

A pair printing mode; pairs should all be displayed using classic s-expressions with dot-deliminated pairs, using indentation to show nesting.

See also PairMode.

Hrse.EQUALS_MODE — Constant

A pair printing mode; pairs should all be displayed using equals-sign-deliminated pairs with implied parentheses, using indentation to show nesting.

See also PairMode.

Hrse.COLON_MODE — Constant

A pair printing mode; pairs should all be displayed using colon-deliminated pairs with implied parentheses, using indentation to show nesting and avoiding parentheses around lists as possible.

See also PairMode.